PDU、母线、机架新更换新表 2024/12/4 jiangjinchi

+++ yudao-module-pdu/yudao-module-pdu-biz/src/main/java/cn/iocoder/yudao/module/pdu/controller/admin/energyconsumption/VO/EleTotalRealtimeRespVO.java
@@ -24,7 +24,7 @@ import lombok.Data;
 public class EleTotalRealtimeRespVO {
     @ExcelProperty("PduId")
     @ExcelIgnore
-    private Long pduId;
+    private Integer pduId;
     @ExcelProperty("所在位置")
     private String address;
     @ExcelProperty("网络地址")
